Create integration test for auto-baseline
